Internal (find-and-replace): Bulk find and replace utils into plugin.
