+++ packages/ckeditor5-find-and-replace/src/findandreplaceediting.js
@@ -8,13 +8,13 @@
  */
 
 import { Plugin } from 'ckeditor5/src/core';
-import { updateFindResultFromRange } from './utils';
 import FindCommand from './findcommand';
 import ReplaceCommand from './replacecommand';
 import ReplaceAllCommand from './replaceallcommand';
 import FindNextCommand from './findnextcommand';
 import FindPreviousCommand from './findpreviouscommand';
 import FindAndReplaceState from './findandreplacestate';
+import FindAndReplaceUtils from './findandreplaceutils';
 
 // eslint-disable-next-line ckeditor5-rules/ckeditor-imports
 import { scrollViewportToShowTarget } from '@ckeditor/ckeditor5-utils/src/dom/scroll';
@@ -26,9 +26,10 @@ import '../theme/findandreplace.css';
 const HIGHLIGHT_CLASS = 'ck-find-result_selected';
 
 // Reacts to document changes in order to update search list.
-function onDocumentChange( results, model, searchCallback ) {
+function onDocumentChange( results, editor, searchCallback ) {
 	const changedNodes = new Set();
 	const removedMarkers = new Set();
+	const model = editor.model;
 
 	const changes = model.document.differ.getChanges();
 
@@ -73,7 +74,8 @@ function onDocumentChange( results, model, searchCallback ) {
 
 	// Run search callback again on updated nodes.
 	changedNodes.forEach( nodeToCheck => {
-		updateFindResultFromRange( model.createRangeOn( nodeToCheck ), model, searchCallback, results );
+		const findAndReplaceUtils = editor.plugins.get( 'FindAndReplaceUtils' );
+		findAndReplaceUtils.updateFindResultFromRange( model.createRangeOn( nodeToCheck ), model, searchCallback, results );
 	} );
 }
 
@@ -83,6 +85,13 @@ function onDocumentChange( results, model, searchCallback ) {
  * @extends module:core/plugin~Plugin
  */
 export default class FindAndReplaceEditing extends Plugin {
+	/**
+	 * @inheritDoc
+	 */
+	static get requires() {
+		return [ FindAndReplaceUtils ];
+	}
+
 	/**
 	 * @inheritDoc
 	 */
@@ -173,7 +182,7 @@ export default class FindAndReplaceEditing extends Plugin {
 		this._activeResults = results;
 
 		// @todo: handle this listener, another copy is in findcommand.js file.
-		this.listenTo( model.document, 'change:data', () => onDocumentChange( this._activeResults, model, findCallback ) );
+		this.listenTo( model.document, 'change:data', () => onDocumentChange( this._activeResults, editor, findCallback ) );
 
 		return this._activeResults;
 	}
